requirements.txt
requests>=2.28.0
PyYAML>=6.0
urllib3>=1.26.0
schedule>=1.2.0
apprise>=1.9.0
python-dotenv>=0.19.0
Flask>=2.3.0
Flask-CORS>=4.0.0
Werkzeug>=2.3.0